@@ -1,4 +1,16 @@
 <?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the FOSOAuthServerBundle package.
+ *
+ * (c) FriendsOfSymfony <http://friendsofsymfony.github.com/>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace FOS\OAuthServerBundle\Tests\DependencyInjection;
 
 use FOS\OAuthServerBundle\DependencyInjection\Configuration;
@@ -6,7 +18,7 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
 use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\Config\Definition\Processor;
 
-class ConfigurationTest extends \PHPUnit_Framework_TestCase
+class ConfigurationTest extends \PHPUnit\Framework\TestCase
 {
     public function testShouldImplementConfigurationInterface()
     {
@@ -14,9 +26,17 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
 
         $this->assertTrue($rc->implementsInterface(ConfigurationInterface::class));
     }
+
     public function testCouldBeConstructedWithoutAnyArguments()
     {
-        new Configuration();
+        try {
+            new Configuration();
+
+            // no exceptions were thrown
+            self::assertTrue(true);
+        } catch (\Exception $exception) {
+            $this->fail($exception->getMessage());
+        }
     }
 
     public function testShouldNotMandatoryServiceIfNotCustomDriverIsUsed()
@@ -33,18 +53,18 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
         ]]);
 
         $this->assertArraySubset([
-            "db_driver" => "orm",
-            "client_class" => "aClientClass",
-            "access_token_class" => "anAccessTokenClass",
-            "refresh_token_class" => "aRefreshTokenClass",
-            "auth_code_class" => "anAuthCodeClass",
-            "service" => [
-                "storage" => "fos_oauth_server.storage.default",
-                "user_provider" => null,
-                "client_manager" => "fos_oauth_server.client_manager.default",
-                "access_token_manager" => "fos_oauth_server.access_token_manager.default",
-                "refresh_token_manager" => "fos_oauth_server.refresh_token_manager.default",
-                "auth_code_manager" => "fos_oauth_server.auth_code_manager.default",
+            'db_driver' => 'orm',
+            'client_class' => 'aClientClass',
+            'access_token_class' => 'anAccessTokenClass',
+            'refresh_token_class' => 'aRefreshTokenClass',
+            'auth_code_class' => 'anAuthCodeClass',
+            'service' => [
+                'storage' => 'fos_oauth_server.storage.default',
+                'user_provider' => null,
+                'client_manager' => 'fos_oauth_server.client_manager.default',
+                'access_token_manager' => 'fos_oauth_server.access_token_manager.default',
+                'refresh_token_manager' => 'fos_oauth_server.refresh_token_manager.default',
+                'auth_code_manager' => 'fos_oauth_server.auth_code_manager.default',
             ],
         ], $config);
     }
@@ -54,7 +74,8 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
         $configuration = new Configuration();
         $processor = new Processor();
 
-        $this->setExpectedException(InvalidConfigurationException::class, 'Invalid configuration for path "fos_oauth_server": The service client_manager must be set explicitly for custom db_driver.');
+        $this->expectException(InvalidConfigurationException::class);
+        $this->expectExceptionMessage('Invalid configuration for path "fos_oauth_server": The service client_manager must be set explicitly for custom db_driver.');
 
         $processor->processConfiguration($configuration, [[
             'db_driver' => 'custom',
@@ -62,7 +83,6 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
             'access_token_class' => 'anAccessTokenClass',
             'refresh_token_class' => 'aRefreshTokenClass',
             'auth_code_class' => 'anAuthCodeClass',
-
         ]]);
     }
 
@@ -71,7 +91,8 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
         $configuration = new Configuration();
         $processor = new Processor();
 
-        $this->setExpectedException(InvalidConfigurationException::class, 'Invalid configuration for path "fos_oauth_server": The service access_token_manager must be set explicitly for custom db_driver.');
+        $this->expectException(InvalidConfigurationException::class);
+        $this->expectExceptionMessage('Invalid configuration for path "fos_oauth_server": The service access_token_manager must be set explicitly for custom db_driver.');
 
         $processor->processConfiguration($configuration, [[
             'db_driver' => 'custom',
@@ -80,8 +101,8 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
             'refresh_token_class' => 'aRefreshTokenClass',
             'auth_code_class' => 'anAuthCodeClass',
             'service' => [
-                'client_manager' => 'a_client_manager_id'
-            ]
+                'client_manager' => 'a_client_manager_id',
+            ],
         ]]);
     }
 
@@ -90,7 +111,8 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
         $configuration = new Configuration();
         $processor = new Processor();
 
-        $this->setExpectedException(InvalidConfigurationException::class, 'Invalid configuration for path "fos_oauth_server": The service refresh_token_manager must be set explicitly for custom db_driver.');
+        $this->expectException(InvalidConfigurationException::class);
+        $this->expectExceptionMessage('Invalid configuration for path "fos_oauth_server": The service refresh_token_manager must be set explicitly for custom db_driver.');
 
         $processor->processConfiguration($configuration, [[
             'db_driver' => 'custom',
@@ -101,7 +123,7 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
             'service' => [
                 'client_manager' => 'a_client_manager_id',
                 'access_token_manager' => 'anId',
-            ]
+            ],
         ]]);
     }
 
@@ -110,7 +132,8 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
         $configuration = new Configuration();
         $processor = new Processor();
 
-        $this->setExpectedException(InvalidConfigurationException::class, 'Invalid configuration for path "fos_oauth_server": The service auth_code_manager must be set explicitly for custom db_driver.');
+        $this->expectException(InvalidConfigurationException::class);
+        $this->expectExceptionMessage('Invalid configuration for path "fos_oauth_server": The service auth_code_manager must be set explicitly for custom db_driver.');
 
         $processor->processConfiguration($configuration, [[
             'db_driver' => 'custom',
@@ -122,8 +145,7 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
                 'client_manager' => 'a_client_manager_id',
                 'access_token_manager' => 'anId',
                 'refresh_token_manager' => 'anId',
-
-            ]
+            ],
         ]]);
     }
 
@@ -143,22 +165,22 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
                 'access_token_manager' => 'an_access_token_manager_id',
                 'refresh_token_manager' => 'a_refresh_token_manager_id',
                 'auth_code_manager' => 'an_auth_code_manager_id',
-            ]
+            ],
         ]]);
 
         $this->assertArraySubset([
-            "db_driver" => "custom",
-            "client_class" => "aClientClass",
-            "access_token_class" => "anAccessTokenClass",
-            "refresh_token_class" => "aRefreshTokenClass",
-            "auth_code_class" => "anAuthCodeClass",
-            "service" => [
-                "storage" => "fos_oauth_server.storage.default",
-                "user_provider" => null,
-                "client_manager" => "a_client_manager_id",
-                "access_token_manager" => "an_access_token_manager_id",
-                "refresh_token_manager" => "a_refresh_token_manager_id",
-                "auth_code_manager" => "an_auth_code_manager_id",
+            'db_driver' => 'custom',
+            'client_class' => 'aClientClass',
+            'access_token_class' => 'anAccessTokenClass',
+            'refresh_token_class' => 'aRefreshTokenClass',
+            'auth_code_class' => 'anAuthCodeClass',
+            'service' => [
+                'storage' => 'fos_oauth_server.storage.default',
+                'user_provider' => null,
+                'client_manager' => 'a_client_manager_id',
+                'access_token_manager' => 'an_access_token_manager_id',
+                'refresh_token_manager' => 'a_refresh_token_manager_id',
+                'auth_code_manager' => 'an_auth_code_manager_id',
             ],
         ], $config);
     }
