@@ -13,9 +13,8 @@ declare(strict_types=1);
 
 namespace FOS\OAuthServerBundle\Command;
 
-use Symfony\Component\Console\Command\Command;
 use FOS\OAuthServerBundle\Model\ClientManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
