@@ -81,17 +81,17 @@ class OAuthProvider implements AuthenticationProviderInterface
             $scope = $accessToken->getScope();
             $user = $accessToken->getUser();
 
-                if (null !== $user) {
-                    try {
-                        $this->userChecker->checkPreAuth($user);
-                    } catch (AccountStatusException $e) {
-                        throw new OAuth2AuthenticateException(Response::HTTP_UNAUTHORIZED,
+            if (null !== $user) {
+                try {
+                    $this->userChecker->checkPreAuth($user);
+                } catch (AccountStatusException $e) {
+                    throw new OAuth2AuthenticateException(Response::HTTP_UNAUTHORIZED,
                             OAuth2::TOKEN_TYPE_BEARER,
                             $this->serverService->getVariable(OAuth2::CONFIG_WWW_REALM),
                             'access_denied',
                             $e->getMessage()
                         );
-                    }
+                }
 
                 $token->setUser($user);
             }
@@ -110,17 +110,17 @@ class OAuthProvider implements AuthenticationProviderInterface
             $token->setAuthenticated(true);
             $token->setToken($tokenString);
 
-                if (null !== $user) {
-                    try {
-                        $this->userChecker->checkPostAuth($user);
-                    } catch (AccountStatusException $e) {
-                        throw new OAuth2AuthenticateException(Response::HTTP_UNAUTHORIZED,
+            if (null !== $user) {
+                try {
+                    $this->userChecker->checkPostAuth($user);
+                } catch (AccountStatusException $e) {
+                    throw new OAuth2AuthenticateException(Response::HTTP_UNAUTHORIZED,
                             OAuth2::TOKEN_TYPE_BEARER,
                             $this->serverService->getVariable(OAuth2::CONFIG_WWW_REALM),
                             'access_denied',
                             $e->getMessage()
                         );
-                    }
+                }
 
                 $token->setUser($user);
             }
